+++ version2.1/php/inventario/inventario.php
@@ -38,7 +38,7 @@ $query = mysqli_query($result,'select * from productos where idproductos != 0 or
 				<td align='right'>$ " . number_format($sumtotal, 0, ",", ".") 	. "</td>
 				<td><a onclick='javascript:abrir(\"editarProductos.php?id=" . $row['idproductos'] . "\")'><span data-tooltip='Editar'>
 				<i class='fa fa-file-text-o nav_icon'></i></spam></a>
-				<a href='eliminarProductos.php?id=" . $row['idproductos'] . "'><span data-tooltip='Eliminar'>
+				<a onClick=\"return confirmar('¿Estas seguro de eliminar?')\" href='eliminarProductos.php?id=" . $row['idproductos'] . "'><span data-tooltip='Eliminar'>
 				<i class='fa icon-off nav-icon'></i></spam></a></td>
 			</tr>";
 
@@ -92,6 +92,16 @@ $html="<!DOCTYPE html>
 	function abrir(url) { 
 	open(url,'','top=100,left=100,width=900,height=500') ; 
 	}
+</script>
+<script>
+function confirmar(texto)
+{
+if (confirm(texto))
+{
+return true;
+}
+else return false;
+}
 </script>
 <!-- //tables -->
 </head>
